+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Deprecated
 
 ### Removed
+- Remove deprecated methods from `TestingFramework` (#285)
 - Remove deprecated parameter from `SalutationSwitcher::translate` (#282)
 - Remove deprecated methods from the `Db` class (#281)
 - Remove non-namespaced Extbase model and repository (#280)